% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OISST_sub_dl.r
\name{OISST_sub_dl}
\alias{OISST_sub_dl}
\title{Download NOAA OISST Data}
\usage{
OISST_sub_dl(
  start,
  end,
  latitude_ext,
  longitude_ext,
  url = "https://coastwatch.pfeg.noaa.gov/erddap/",
  dataset = "ncdcOisst21Agg_LonPM180"
)
}
\arguments{
\item{start}{Start date. Must be formatted with \code{as.Date()}}

\item{end}{End date. Must be formatted with \code{as.Date()}}

\item{latitude_ext}{Vector of two coordinates with latitude extent.}

\item{longitude_ext}{Vector of two coordinates with longitude extent}

\item{url}{URL of ERDDAP server. Defaults to https://coastwatch.pfeg.noaa.gov/erddap}

\item{dataset}{Dataset used on ERDDAP server. Defaults to ncdcOisst21Agg_LonPM180 See https://coastwatch.pfeg.noaa.gov/erddap/search/index.html?page=1&itemsPerPage=1000&searchFor=OISST for more.}
}
\value{
A data frame of latitude, longitude, date, and temperature
}
\description{
Download NOAA OISST Data
}
\examples{
\dontrun{
library(ggplot2)

temp <- OISST_sub_dl(
  start = as.Date("1982-01-01"),
  end = as.Date("1989-12-31"),
  latitude_ext = c(-40, -35),
  longitude_ext = c(15, 21)
)

  ggplot(data = temp,
    aes(x = lon, y = lat)) +
    geom_tile(aes(fill = temp)) +
    scale_fill_viridis_c() +
    coord_quickmap(expand = F) +
    labs(x = NULL, y = NULL, fill = "SST (°C)") +
    theme(legend.position = "bottom")
}
}
